requirements.txt
# --- Core AI & Framework ---
torch
torchaudio
torchvision
numpy
pillow
python-dotenv
prompt-toolkit
colorama

# --- Vision & Sensors (Area strictly preserved) ---
ultralytics
opencv-python
psutil
paho-mqtt
roboflow

# --- RAG & Data Engine ---
langchain
langchain-community
langchain-text-splitters
faiss-cpu
sentence-transformers
ollama
pypdf
python-docx
python-pptx
olefile
beautifulsoup4
requests

# --- Voice Processing (STT & TTS) ---
faster-whisper
pyttsx3
pygame
pyaudio
